+++ src/components/admin/ItemDelete.vue
@@ -7,22 +7,19 @@
 </template>
 
 <script>
-import axios from "axios"
-
 export default {
   methods: {
     goBack() {
       document.location.href = "/collection/" + this.$route.params.cid + "/manage/"
     },
     deleteItem() {
-      // API call
-      axios.delete(import.meta.env.VITE_API_URL + "/collections/" + this.$route.params.cid + "/items/" + this.$route.params.id, this.$apiConfig())
-      .then(() => {
-        this.goBack()
-      })
-      .catch(() => {
-        alert('Error')
-      })
+      this.$apiDelete("collections/" + this.$route.params.cid + "/items/" + this.$route.params.id)
+        .then(() => {
+          this.goBack()
+        })
+        .catch(() => {
+          alert("Error")
+        })
     }
   }
 }