+++ src/components/user/ChangePassword.vue
@@ -1,33 +1,29 @@
 <template>
   <form @submit="changePassword">
     <div v-if="!resetToken" class="form-floating">
-        <input v-model="oldPassword" type="password" class="form-control" id="oldPassword" placeholder="Current password" required>
-        <label for="oldPassword">Current password</label>
+      <input v-model="oldPassword" type="password" class="form-control" id="oldPassword" placeholder="Current password" required>
+      <label for="oldPassword">Current password</label>
     </div>
     <div class="form-floating">
-        <input v-model="newPassword" type="password" class="form-control" id="newPassword" placeholder="New password" required>
-        <label for="newPassword">New password</label>
+      <input v-model="newPassword" type="password" class="form-control" id="newPassword" placeholder="New password" required>
+      <label for="newPassword">New password</label>
     </div>
     <div class="form-floating">
-        <input v-model="confirmPassword" type="password" class="form-control" id="confirmPassword" placeholder="Confirm new password" required>
-        <label for="confirmPassword">Confirm new password</label>
+      <input v-model="confirmPassword" type="password" class="form-control" id="confirmPassword" placeholder="Confirm new password" required>
+      <label for="confirmPassword">Confirm new password</label>
     </div>
     <button class="w-100 btn btn-lg btn-primary" type="submit" :disabled="$demoMode()">Change password</button>
     <div v-if="success" class="alert alert-success" role="alert">
-        Password changed
+      Password changed
     </div>
     <div v-if="notSimilar" class="alert alert-danger" role="alert">
-        Passwords not equals!
+      Passwords not equals!
     </div>
-    <div v-if="error" class="alert alert-danger" role="alert">
-        Failed!
-    </div>
+    <div v-if="error" class="alert alert-danger" role="alert">Failed!</div>
   </form>
 </template>
 
 <script>
-import axios from "axios"
-
 export default {
   props: {
     resetToken: {
@@ -66,7 +62,7 @@ export default {
         data.password = this.oldPassword
       }
 
-      axios.post(import.meta.env.VITE_API_URL + url, data, this.$apiConfig())
+      this.$apiPost(url, data)
         .then(() => {
           this.error = false
           this.success = true