+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix crash when setting searchable property
 - Display thumbnails instead of full images
 - Many design improvements
+- Various code improvements
 
 # 1.0 RC 1 (2022-05-28)
 - Items sorting and filters are now working properly